requirements.txt
click>=8.1.7,<9
PyYAML>=6.0.1,<7
mysql-connector-python>=9.0.0,<10
pytest>=8.3.2,<9
pytest-mock>=3.14.0,<4
